==> encoding/01_encoding_proc_envelope.py <==
# -*- coding: utf-8 -*-

#!/usr/bin/env python
import numpy as np
import glob
import os.path
import json
import re
import scipy.io.wavfile as wav
from scipy import stats
import librosa as lbr 
from scipy.signal import hilbert
import librosa.display
import pickle
import pywt
import config_analysis

def get_mel_spectrogram(filename, log=True, sr=44100, hop_length=512, **kwargs):
#    '''Returns the (log) Mel spectrogram of a given wav file, the sampling rate
#    of that spectrogram and names of the frequencies in the Mel spectrogram
#
#    Parameters
#    ----------
#    filename : str, path to wav file to be converted
#    sr : int, sampling rate for wav file
#         if this differs from actual sampling rate in wav it will be resampled
#    log : bool, indicates if log mel spectrogram will be returned
#    kwargs : additional keyword arguments that will be
#             transferred to librosa's melspectrogram function
#
#    Returns
#    -------
#    a tuple consisting of the Melspectrogram of shape (time, mels), the 
#    repetition time in seconds, and the frequencies of the Mel filters in Hertz 
#    '''
    wav, _ = lbr.load(filename, sr=sr)
    melspecgrams = lbr.feature.melspectrogram(y=wav, sr=sr, hop_length=hop_length,
                                              **kwargs)
    if log:
        melspecgrams[np.isclose(melspecgrams, 0)] = np.finfo(melspecgrams.dtype).eps
        melspecgrams = np.log(melspecgrams)
    log_dict = {True: 'Log ', False: ''}
    freqs = lbr.core.mel_frequencies(
            **{param: kwargs[param] for param in ['n_mels', 'fmin', 'fmax', 'htk']
                if param in kwargs})
    freqs = ['{0:.0f} Hz ({1}Mel)'.format(freq, log_dict[log]) for freq in freqs]
    return melspecgrams.T, sr / hop_length, freqs

if __name__ == "__main__":
    # =============================================================================
    # Load data 
    # =============================================================================
    repository = os.path.join(config_analysis.project_root, "gaudie_audio_validation")

    save_path = os.path.join(repository, 'derivatives', 'encoding', 'encoding_mel_hilb')
    if not os.path.exists("{}".format(save_path)):
        print('creating path for saving')
        os.makedirs("{}".format(save_path))
    save_path_meta = os.path.join(save_path, 'meta_data')
    if not os.path.exists("{}".format(save_path_meta)):
        print('creating path for saving meta_data')
        os.makedirs("{}".format(save_path_meta))    
    wav_files = glob.glob(os.path.join(repository, 'stimulus_material', "*.wav"))
    dict_sr = {}
    dict_freq = {}
    
    for wav_file in wav_files:
        print("Converting ",wav_file)
        
        # sig2, Fs = lbr.load(wav_file, sr=None, mono=True)
        rate, sig = wav.read(wav_file)
        if len(sig.shape) > 1:
               sig = np.mean(sig,axis=1) # convert a WAV from stereo to mono
        ## set parameters ##
        #rate        = 44100                     # sampling rate
        winlen      = int(np.rint(rate*0.025))  # 1102 Window length std 0.025s
        overlap     = 0
        hoplen      = winlen-overlap            # 661 hop_length
        nfft        = winlen    # standard is = winlen = 1102 ... winlen*2 = 2204 ... nfft = the FFT size. Default for speech processing is 512.
        nmel        = 32       # n = the number of cepstrum to return = the number of filters in the filterbank
        lowfreq     = 100       # lowest band edge of mel filters. In Hz
        highfreq    = 8000      # highest band edge of mel filters. In Hz
        noay        = 3         # subplot: y-dimension
        noax        = 1         # subplot: x-dimension
        # foursec     = int(np.rint(rate*4.0)) # 4 seconds
        # start_time  = 0
    
        config = {"n_fft":nfft, "sr":rate, "win_length":winlen, 
                  "hop_length":hoplen, "n_mels":nmel, "fmax":highfreq, 
                  "fmin":lowfreq}  
        melspec, sr_spec, freqs = get_mel_spectrogram(wav_file, **config)
        zscore_melspec= stats.zscore(melspec, axis=0)
        zband_envelope= np.abs(hilbert(zscore_melspec, axis=0))
        band_envelope = np.abs(hilbert(melspec, axis=0))
        
        log_mean_zband_envelope = np.log1p(np.mean(zband_envelope,axis=1))
        log_mean_band_envelope = np.log1p(np.mean(band_envelope,axis=1))
        
        mean_zband_envelope = np.mean(zband_envelope,axis=1)
        mean_band_envelope = np.mean(band_envelope,axis=1)
        
        outfile_base = os.path.basename(wav_file).split('.')[0][17:] 
        dict_sr[outfile_base] = sr_spec
        dict_freq[outfile_base] = freqs
        with open(os.path.join(save_path,'envelope_hilb_z_score_'+ outfile_base +'.pkl'), 'wb') as f:
            pickle.dump(mean_zband_envelope, f)
        with open(os.path.join(save_path,'envelope_hilb_'+ outfile_base +'.pkl'), 'wb') as f:
            pickle.dump(mean_band_envelope, f)
            
    with open(os.path.join(save_path_meta,'dict_sr.pkl'), 'wb') as f:
        pickle.dump(dict_sr, f)
    with open(os.path.join(save_path_meta,'dict_freq.pkl'), 'wb') as f:
        pickle.dump(dict_freq, f)


    #tsv_file = os.path.join(subses_output_folder, outfile_base+'.tsv.gz')
    #json_file = os.path.join(subses_output_folder, outfile_base+'.json')
    #png_file = os.path.join(subses_output_folder, outfile_base+'.png')
                
        
    #np.savetxt(tsv_file, mean_zband_envelope, delimiter='\t')
    #metadata = {'SamplingFrequency': sr_spec, 'StartTime': 0,
    #           'Columns': [0]}
    #with open(json_file, 'w+') as fp:
    #    json.dump(metadata, fp)
